Refactored ModelBasePath into separate file
