+++ dbtwiz/model/validate.py
@@ -8,7 +8,7 @@ from ruamel.yaml import YAML
 
 from dbtwiz.config.project import ProjectConfig
 from dbtwiz.dbt.manifest import Manifest
-from dbtwiz.dbt.project import ModelBasePath
+from dbtwiz.dbt.model import ModelBasePath
 from dbtwiz.gcp.auth import ensure_app_default_auth
 from dbtwiz.gcp.bigquery import BigQueryClient
 from dbtwiz.helpers.logger import status, warn
